==> 4.4_candidate-tf-analysis.Rmd <==
---
title: "TF Analysis"
author: "Brennan McDonald"
date: "2024-04-03"
output: html_document
---

# Overview

The following script provides the code needed to analyze the expression profiles (using the outputs from the Mfuzz pipeline) for the transcription factors in the list generated by the "2.2_tf-list-curation.Rmd" script. This data will be used to generate a candidate list of TFs that may regulate adult rudiment development in He. The plots generated by this script are used in Figure 8.


```{r}
# Load packages
library(tidyverse)
library(Seurat)
library(scCustomize)
```


```{r}
# Load the Mfuzz analysis object
load(file = "data-files/mfuzz_object.Rda")

# Load the Seurat object
CreateSeuratObject(load(
  file = "HE_6-60hpf_integrated-SCT6k_195dim.Rda", 
                        verbose = T))
```


# Assign Mfuzz expression profiles to TF genes

We matched the expression profile assignments from the Mfuzz analysis with the TF genes from my curated list (produced by the "2.2_tf-list-curation.Rmd" script).

```{r}
# List of all genes with MFuzz cluster assignments
mfuzz_genes <- read.csv("data-files/gene_cluster_assignments.csv")

# List of candidate transcription facotrs
tfs <- read.csv("data-files/devens-and-go-term_tf-list.csv")


# Generate gene codes to use to lookup the corresponding entries in the Mfuzz 
# table and Seurat object
tfs <- tfs %>% 
  mutate(new_gid = str_replace_all(GID, pattern = "_", replace = "-"),
         new_spu_name = str_replace_all(SPU_common_name, 
                                        pattern = "_", 
                                        replace = "-")) %>% 
  mutate(seurat_code = paste(new_gid, new_spu_name, sep = ":")) %>% 
  mutate(mfuzz_code = str_replace_all(seurat_code, pattern = ":|-|/|_", replace = ".")) %>% 
  dplyr::select(-new_gid, -new_spu_name)


# This identifies the genes from the TF list that are included in the Mfuzz output
# An inner_join() only keeps observations from x that have a matching key in y.
mfuzz_tfs <- inner_join(x = as_tibble(mfuzz_genes), y = as_tibble(tfs),
           by = join_by(code == mfuzz_code),
           na_matches = "never") %>% 
  dplyr::select(-GID.y) %>% 
  rename(GID = GID.x)
```


This code then classifies whether the cluster assignments for each TF gene fall into clusters that are part of the "High Early", "High Middle", or "High Late" group (from the output of the heatmap k-means analysis in "mfuzz-pipeline.Rmd").

```{r}
# sum number of TF genes in each cluster depending on the shape of their expression profiles
exp_early <- c(3,5,8)
exp_middle <- c(1,6,7)
exp_late <- c(2,4,9)
mfuzz_tfs$high_expression <- rep(NA, nrow(mfuzz_tfs))
for (i in 1:nrow(mfuzz_tfs)) {
  if (mfuzz_tfs[i,"clusters"] %in% exp_early) {
    mfuzz_tfs[i,"high_expression"] <- "High Early"
  } else if (mfuzz_tfs[i,"clusters"] %in% exp_middle) {
    mfuzz_tfs[i,"high_expression"] <- "High Middle"
  } else if (mfuzz_tfs[i,"clusters"] %in% exp_late) {
    mfuzz_tfs[i,"high_expression"] <- "High Late"
  }
}

write.csv(mfuzz_tfs, "data-files/mfuzz_tfs.csv", row.names = F)
```


# Identify candidate TFs regulating adult rudiment development

The TF list is filtered for genes with "High Late" cluster profile assignments.

```{r}
# TFs that were found in the Mfuzz object
mfuzz_tfs <- read_csv("data-files/mfuzz_tfs.csv")

# Filter for TFs that have late peaks to their expression
late_tf <- mfuzz_tfs %>% 
  filter(high_expression == "High Late")

mfuzz_tfs %>% 
  group_by(high_expression) %>% 
  summarize(n = n())
```


We then filter the late-expressed TFs for genes that show strong localization to the cell types in the He Seurat object that are found in the adult rudiment (Rudiment Ectoderm (Verified), Rudiment Ectoderm (Putative), Posterior Ectoderm).

```{r}
# For each of the TFs, extract their fold-change enrichment in each of the 
# 3 cell types of interest relative to all other cell types in the Seurat object
types = c("Rudiment Ectoderm (Putative)", "Rudiment Ectoderm (Verified)", 
          "Posterior Ectoderm")
tfs <- tibble(Gene = c(NA),
              cell_type = c(NA),
              avg_log2FC = c(NA),
              pct.1 = c(NA),
              pct.2 = c(NA))
for (gene in late_tf$seurat_code) {
  for (type in types) {
    fc <- FoldChange(he.integrated, features = gene, ident.1 = type)
    tfs <- add_row(tfs, 
                   Gene = gene,
                   cell_type = type,
                   avg_log2FC = fc[1,1],
                   pct.1 = fc[1,2],
                   pct.2 = fc[1,3])
  }
}
tfs <- tfs[-c(1),]

# Filter for TFs that are strongly restricted to the cell types of interest
# They should be expressed in fewer than 10% of the cells in the other cell types
# (pct.2 < 0.1), more than 5% of cells in any rudiment cell type (pct.1 > 0.05),
# and have a log2 fold change greater than 0.9 in any rudiment cell type 
# (avg_log2FC > 0.9)
candidate_tfs <- tfs %>% 
  filter(pct.2 < 0.1) %>% 
  filter(pct.1 > 0.05) %>% 
  filter(avg_log2FC > 0.9) %>% 
  group_by(cell_type) %>% 
  slice_max(n = 50, order_by = avg_log2FC)

# This represents the list of 33 unique candidate transcription factors for 
# regulating adult rudiment development
write.csv(candidate_tfs, "data-files/candidate-tf-list.csv",
          row.names = F)

# Retain the unfiltered TF list for plotting. If a TF met the conditions described
# above, the choose variable will get a value "select" for plotting purposes
tfs <- tfs %>%
  mutate(choose = if_else(avg_log2FC > 0.9 & pct.1 > 0.05 & pct.2 < 0.1, "Select", "No"))
```


The following code generates the plot in Figure 8 that represents the filtration process for the TF list. All of the initial TFs are plotted based on their pct.2 and pct.1 in the rudiment cell types, and they are highlighted in red if they meet the criteria described above for being a candidate TF for adult rudiment development.

```{r}
# Plot all the late stage TFs, and highlight the TFs that meet the selection criteria
tfs %>% 
  ggplot() + 
  geom_point(mapping = aes(x = pct.1, y = pct.2, color = avg_log2FC), size = 0.9) + 
  #scale_color_viridis_c(option = "magma")
  scale_color_gradient(low = "lightblue", high = "darkblue",
                       name = "Average\nLog2(FC)") +
  geom_point(data = tfs[tfs$choose == "Select",], 
             aes(x = pct.1, y = pct.2), color = "red", size = 1.1) +
  theme_bw() +
  geom_line(data = data.frame(x = c(0.05,Inf), y = c(0.1, 0.1)), 
            aes(x = x, y = y), linetype = 2) +
  geom_line(data = data.frame(x = c(0.05,0.05), y = c(-Inf, 0.1)), 
            aes(x = x, y = y), linetype = 2) +
  labs(x = "Percent in Cell Type of Interest",
       y = "Percent in Other Cell Types")
```


## Export featureplots for figure 8

These are featare plots for the 4 candidate TFs shown in Figure 8.

```{r}
# Export as 6x8.5 for the feature plot of Msx in Fig. 8
FeaturePlot_scCustom(he.integrated, 
                     features = "HER-9194.t1:Sp-Msx",
                     colors_use = viridis_dark_high,
                     pt.size = 0.3,
                     order = T
                     )

# Export as 6x8.5 for the feature plot of Giant in Fig. 8
FeaturePlot_scCustom(he.integrated, 
                     features = "HER-17334.t1:Sp-Giant",
                     colors_use = viridis_dark_high,
                     pt.size = 0.3,
                     order = T
                     )

# Export as 6x8.5 for the feature plot of Tll in Fig. 8
FeaturePlot_scCustom(he.integrated, 
                     features = "HER-35851.t1:Sp-Tll",
                     colors_use = viridis_dark_high,
                     pt.size = 0.3,
                     order = T
                     )

# Export as 6x8.5 for the feature plot of Lox in Fig. 8
FeaturePlot_scCustom(he.integrated, 
                     features = "HER-48747.t1:Sp-Lox",
                     colors_use = viridis_dark_high,
                     pt.size = 0.3,
                     order = T
                     )
```
